bulk update should not be affected by hidden rows
Some studios manage their shows with excel. In an excel, every
shot in the show is exists. They sometimes give us the excel so we
can stay on the same page. In the excel they sometimes hide rows
unrelated with the direction. If we directly use the excel
to update properties, it will replace every data we have.
Which can be disaster. Let's don't be affected by hidden rows.

Close #42