service: now search SELECT FROM entries with JOIN-ing properties
It causes a problem that the entries not having any property
excluded from the search.

The old code worked well until Search started to check
their sub entry properties.

This bug was hard to understand as in real world examples,
it looked like it had no problem. Opposed to the test results.
Because every entry usually have some properties in real world.