implement archiving
Number of shows that are done but cannot be deleted has been grown.
They are needed to check which artist do which work.
But they makes user hard to keep eyes on working projects,
also affect to search speed.

Archieved entries will not be shown in entry list or search,
unless explicitly told.