don't redirect to /login but show login page in that url
redirect to /login makes uncomfortable situation when a user shares
url but other user lost the url when they didn't login forge already.
Of course the user can input the url again, but it is cumbersome.

This problem makes big trouble in Naverworks app which forces to
use in-app browser that google doesn't allow login.