service: add property_owner global and use it for delegating properties
there are cases where a user without having permission to write
to an entry needs to change some properties of the entry.

for example, an artist cannot change shot properties but
a matchmover should be able to change the related shot properties
like undistort plate size property.

so I add mechanism of property owner, who owns delegated properties
but not other properties on the entry.

property owner still cannot change or see the entry (and property)
if they don't have read access on the entry.